requirements.txt
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
TgCrypto==1.2.5
tqdm==4.66.5
